Rename filename webpacker_test_helper => test_helper
