+++ test/configuration_test.rb
@@ -1,4 +1,4 @@
-require "webpacker_test_helper"
+require "test_helper"
 
 class ConfigurationTest < Webpacker::Test
   def test_source_path
