+++ test/helper_test.rb
@@ -1,4 +1,4 @@
-require "webpacker_test_helper"
+require "test_helper"
 
 class HelperTest < ActionView::TestCase
   def setup
